- Backend update from Raze.

+++ src/common/utility/cmdlib.cpp
@@ -986,7 +986,7 @@ bool IsAbsPath(const char *name)
     if (IsSeperator(name[0])) return true;
 #ifdef _WIN32
     /* [A-Za-z]: (for Windows) */
-    if (isalpha(name[0]) && name[1] == ':')    return true;
+    if (isalpha((uint8_t)name[0]) && name[1] == ':')    return true;
 #endif /* _WIN32 */
     return 0;
 }
